Update to larval/lumen 5.6 and unique tests for both
