+++ .travis.yml
@@ -1,8 +1,7 @@
 language: php
 php:
-  - "7.0"
-  - "7.1"
-  - "hhvm"
+  - "7.13"
+  - "7.2"
 
 before_script:
   - curl -sS https://getcomposer.org/installer | php -- --filename=composer
@@ -14,10 +13,7 @@ script:
 
 after_script:
   - php vendor/bin/codacycoverage clover build/logs/clover.xml
-
-matrix:
-     allow_failures:
-         - php: "hhvm"
+  
 branches:
   only:
     - master
