--- tests/BaseTestCase.php
@@ -1,59 +1,0 @@
-<?php
-
-namespace Askedio\Tests;
-
-use Illuminate\Filesystem\Filesystem;
-
-class BaseTestCase extends \Illuminate\Foundation\Testing\TestCase
-{
-    /**
-     * Setup DB before each test.
-     */
-    public function setUp()
-    {
-        parent::setUp();
-
-        $this->app['config']->set('database.default', 'sqlite');
-        $this->app['config']->set('database.connections.sqlite.database', ':memory:');
-        $this->app['config']->set('app.url', 'http://localhost/');
-        $this->app['config']->set('app.debug', false);
-        $this->app['config']->set('app.key', env('APP_KEY', '1234567890123456'));
-        $this->app['config']->set('app.cipher', 'AES-128-CBC');
-
-        $this->app->boot();
-
-        $this->migrate();
-    }
-
-    /**
-     * Run package database migrations.
-     */
-    public function migrate()
-    {
-        $fileSystem = new Filesystem();
-        $classFinder = new ClassFinder();
-
-        foreach ($fileSystem->files(__DIR__.'/App/database/migrations') as $file) {
-            $fileSystem->requireOnce($file);
-            $migrationClass = $classFinder->findClass($file);
-            (new $migrationClass())->down();
-            (new $migrationClass())->up();
-        }
-    }
-
-    /**
-     * Boots the application.
-     *
-     * @return \Illuminate\Foundation\Application
-     */
-    public function createApplication()
-    {
-        $app = require __DIR__.'/../vendor/laravel/laravel/bootstrap/app.php';
-
-        $app->make(\Illuminate\Contracts\Console\Kernel::class)->bootstrap();
-
-        $app->register(\Askedio\SoftCascade\Providers\GenericServiceProvider::class);
-
-        return $app;
-    }
-}
