+++ tests/TestCase.php
@@ -1,0 +1,58 @@
+<?php
+
+namespace Askedio\Tests;
+
+use Orchestra\Testbench\TestCase as BaseTestCase;
+use Illuminate\Filesystem\Filesystem;
+
+class TestCase extends BaseTestCase
+{
+    /**
+     * Get package providers.
+     *
+     * @param \Illuminate\Foundation\Application $app
+     *
+     * @return array
+     */
+    protected function getPackageProviders($app)
+    {
+        return [
+            \Askedio\SoftCascade\Providers\GenericServiceProvider::class,
+        ];
+    }
+
+    /**
+     * Define environment setup.
+     *
+     * @param Illuminate\Foundation\Application $app
+     *
+     * @return void
+     */
+    public function getEnvironmentSetUp($app)
+    {
+        $app['config']->set('database.default', 'sqlite');
+        $app['config']->set('database.connections.sqlite.database', ':memory:');
+        $app['config']->set('app.url', 'http://localhost/');
+        $app['config']->set('app.debug', false);
+        $app['config']->set('app.key', env('APP_KEY', '1234567890123456'));
+        $app['config']->set('app.cipher', 'AES-128-CBC');
+
+        $this->migrate();
+    }
+
+    /**
+     * Run package database migrations.
+     */
+    public function migrate()
+    {
+        $fileSystem = new Filesystem();
+        $classFinder = new ClassFinder();
+
+        foreach ($fileSystem->files(__DIR__.'/App/database/migrations') as $file) {
+            $fileSystem->requireOnce($file);
+            $migrationClass = $classFinder->findClass($file);
+            (new $migrationClass())->down();
+            (new $migrationClass())->up();
+        }
+    }
+}
